Dev v1.3

+++ app/src/main/java/com/breadwallet/tools/sqlite/BRSQLiteHelper.java
@@ -55,7 +55,7 @@ public class BRSQLiteHelper extends SQLiteOpenHelper {
     }
 
     public static final String DATABASE_NAME = "breadwallet.db";
-    private static final int DATABASE_VERSION = 16;
+    private static final int DATABASE_VERSION = 17;
 
     /**
      * DID author table
@@ -218,7 +218,11 @@ public class BRSQLiteHelper extends SQLiteOpenHelper {
     @Override
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
 
-        db.execSQL("DROP TABLE IF EXISTS " + ELA_TX_TABLE_NAME);
+        if((newVersion!=oldVersion) && newVersion==17) {
+            db.execSQL("DROP TABLE IF EXISTS " + ELA_TX_TABLE_NAME);
+            db.execSQL(ELA_TX_DATABASE_CREATE);
+        }
+
         if (/*oldVersion < 13 && (newVersion >= 13)*/ newVersion==16) {
             boolean migrationNeeded = !tableExists(MB_TABLE_NAME, db);
             onCreate(db); //create new db tables
